task-daily-service/requirements.txt
langchain
langchain_community
langchain_core
requests
beautifulsoup4
lxml
cassandra-driver
langchain_google_genai
holidays